Move content source to a subdirectory.
This commit moves a great deal of the content and Gitbook operations to
a subdirectory.  This is so that when a release occurs, the output
doesn't neighbor the actual workspace.

One major consequence of this is that in order to build the Gitbook you
have to change directories in order to do it, so this commit also
introduces 4 convenience scripts:

`npm run serve` moves into the correct directory for you and serves the
Gitbook in Live Reload mode so you can review the site as you work.

`npm run build` builds a fresh copy of the Gitbook site.

`npm run dist` copies the contents of the Gitbook build to the root
directory, so Github Pages have something to host.

`npm run release` a bundling script that combines `build` and `dist`.