Add Supergiant logo and Doppio One font
In order to bring styles together across Supergiant projects, this
commit ads a logo to the top of the table of contents, a "you are here"
pointer to the active book entry, and styles headlines using Doppio
One, a.k.a. the Supergiant typeface.

In order to reduce confusion between linked items, auto-linked Glossary
items are now text-colored with a dotted underline. Normal links are
unchanged.