Add Contributor Docs
In order to allow others to ammend our code we need license agreemtns,
instructions, and a code of conduct. This commit adds all three.

Beccause there are legal consequences if, say, an employee of a
business contributes code to a Supergiant repo (that company could
claim ownership of what was contributed), this commit adds both an
Individual Contributor License Agreement and a Entity Contributor
License Agreement from harmonyagreements.org -- the same source
Canonical uses, which have stood up in court.

These will set us on the right foot, in case either an individual or a
company wants to get in on the love.