Add Disqus comment box to docs pages
In order to allow community questions and engagement where helpful,
this commit adds a Discus comment box to the footer of each page.

In order to allow easier linking to topics within the documentation,
this commit also adds GitHub-like anchors to H-titles. The hope
behind this is that sections may be more easily referenced and linked
by other authors and from within the documentation, itself.