Add Google Analytics Beacon to root README
In order to see how many anonymous users are interested in our
documentation, this commit ads a Google Analytics beacon to the
root README.md, visible only from github.com. This beacon is not
rendered as part of the reference GitBook, so tracking only occurs
from the github.com domain.