Edit contributor agreements for e-signatures
Contributor forms updated from Harmony to Apache CLA verbiage.

Contributor forms should be e-signable, so this commit removes verbaige
for mailing or emailing, and it points interested would-be contributors
to a Supergiant.io friendly URL.

The friendly URL is a single point in the process that may be used to
301 users to any e-signature form handler the Supergiant team may use
in the future.