Remove empty, misleading Getting Started page
As noted in #13, the Getting Started page is full of blank links. What
started as a documentation page became our intro video script and then
was never executed. This commit removes the page of dead links.

This does not fix the issue report, which should remain open until the
section contains proper getting started documentation.